+++ lib/bright_serializer/inflector.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class Inflector
-  CAMEL_REGEX = /(_[a-zA-Z])/.freeze
+  CAMEL_REGEX = /(_.)/.freeze
   UNDERSCORE_REGEX = /(.)([A-Z])/.freeze
 
   class << self
@@ -9,13 +9,19 @@ class Inflector
       return term.capitalize! unless term.include?('_')
 
       term.capitalize!
-      term.gsub!(CAMEL_REGEX) { Regexp.last_match(1).delete('_').upcase! }
+      term.gsub!(CAMEL_REGEX) do
+        match = Regexp.last_match(1).delete('_')
+        match.upcase! || match
+      end
     end
 
     def camel_lower(term)
       return term unless term.include?('_')
 
-      term.gsub!(CAMEL_REGEX) { Regexp.last_match(1).delete('_').upcase! }
+      term.gsub!(CAMEL_REGEX) do
+        match = Regexp.last_match(1).delete('_')
+        match.upcase! || match
+      end
     end
 
     def underscore(camel_cased_word)