+++ spec/lib/bright_serializer/serializer_spec.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require_relative '../../share/user'
+require_relative '../../share/serializers'
 
 RSpec.describe BrightSerializer::Serializer do
   describe '#serialize' do
@@ -140,6 +141,52 @@ RSpec.describe BrightSerializer::Serializer do
       it 'serialize all 3 attributes' do
         expect(child_class.new(user).to_hash).to eq(result)
       end
+    end
+  end
+
+  describe '#has_one' do
+    let(:serializer_class) do
+      Class.new do
+        include BrightSerializer::Serializer
+        attributes :first_name, :last_name
+        has_many :friends, class_name: 'FriendSerializer'
+        has_one :best_friend, class_name: 'FriendSerializer'
+        has_one :second_best_friend, class_name: 'FriendSerializer', if: proc { false }
+        belongs_to :third_best_friend, class_name: 'FriendSerializer', params: { prefix: 'Mr ' }
+      end
+    end
+
+    let(:user) do
+      user = User.new
+      def user.best_friend
+        @best_friend ||= User.new
+      end
+
+      def user.third_best_friend
+        @third_best_friend ||= User.new
+      end
+      user.friends = [User.new, User.new]
+      user
+    end
+
+    let(:expected_result) do
+      {
+        first_name: user.first_name,
+        last_name: user.last_name,
+        friends: [
+          { first_name: user.friends[0].first_name },
+          { first_name: user.friends[1].first_name }
+        ],
+        best_friend: { first_name: user.best_friend.first_name },
+        third_best_friend: {
+          first_name: user.third_best_friend.first_name,
+          last_name: "Mr #{user.third_best_friend.last_name}"
+        }
+      }
+    end
+
+    it 'serializer has_many friends' do
+      expect(serializer_class.new(user).serializable_hash).to eq expected_result
     end
   end
 end